main.c help.h showsuper.c
So I created a terminal main.c, which should be what he is looking for with input. I mounted the imagefile inside of main.c, didn't seem like I needed a separate file for that. I also defined help in a header file, not sure if anyone wants to change that and do it another way, but it just displays all available help commands. I also uploaded my showsuper.c file, which doesn't seem to be fully working yet, but I am working on that.